Update umbraco-marketplace-jumoo.translationmanager.openai.json
Hi Kevin - we're about to introduce a new category on the Marketplace for "Artificial Intelligence" and from reading the details of your package, I think it would make sense to list it in this category - as a secondary one alongside the existing main category.

If you are in agreement, I've raised this PR to do that.  If accepted and merged, your package when synchronised will be listed in the new category once it's published.

Thanks for your support of the Umbraco community with your package.